Update GROMACS extension name
The LigParGen webserver has updated the name of the GROMACS topology to .top, not .itp, and this change is simply making the documentation clear that it still works with the new extension. The name of a variety of variable and functions still uses 'itp' internally, but this doesn't effect the running of the script, and so they will only be changed the next time the script receives an update to the code itself.